Added gui inspector

+++ Source/Editor/Tools/IEditorTool.cpp
@@ -1,4 +1,4 @@
 
 #include "IEditorTool.h"
 std::vector<uint32_t> IEditorTool::mySelectedObjects;
-uint8_t IEditorTool::mySelectedPartID;
+uint32_t IEditorTool::mySelectedPartID;
